+++ packages/ingest/S3ProviderClient.js
@@ -64,8 +64,8 @@ class S3ProviderClient {
    * @param {string} sourceKey - the full path to the remote file to be fetched
    * @param {string} destinationBucket - destination s3 bucket of the file
    * @param {string} destinationKey - destination s3 key of the file
-   * @returns {Promise.<{ s3uri, etag }>} promise resolving to an object
-   *   containing the s3 uri and etag of the destination file
+   * @returns {Promise.<{ s3uri: string, etag: string }>} an object containing
+   *    the S3 URI and ETag of the destination file
    */
   async sync(sourceKey, destinationBucket, destinationKey) {
     try {